feat(server): create for migration job
